+++ agents-run-api/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @inkeep/agents-run-api
 
+## 0.29.2
+
+### Patch Changes
+
+- @inkeep/agents-core@0.29.2
+
+## 0.29.1
+
+### Patch Changes
+
+- f2ac869: upgrade docs
+- 37e50a6: fix mcp headers with context config
+- 65f4b1a: remove builtin time variables from context
+- Updated dependencies [f2ac869]
+- Updated dependencies [37e50a6]
+- Updated dependencies [65f4b1a]
+  - @inkeep/agents-core@0.29.1
+
 ## 0.29.0
 
 ### Patch Changes